saves spectrum for each second

+++ Pickler.py
@@ -1,10 +1,10 @@
 __author__ = 'lukestack'
 
 import os
-from scipy.signal import decimate
+from scipy.signal import resample
 from scipy.io.wavfile import read
 import matplotlib
-matplotlib.use('Agg')
+# matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import tempfile
 from pydub import AudioSegment
@@ -13,9 +13,9 @@ import numpy as np
 import Dates
 
 
-fs = 1225
-nfft = 1024
-noverlap = 512
+fs = 2048
+nfft = 4096
+noverlap = 2048
 
 
 def get_data(path):
@@ -45,25 +45,47 @@ def combine_specgrams(data1, data2):
     return np.hstack((data1, data2))
 
 
-def save_specgram_pkl(data, title=None, name=None, show=False):
-    if show:
-        fig = plt.figure()
-        ax = fig.add_subplot(111)
-        ax.specgram(data,  pad_to=nfft, NFFT=nfft, noverlap=noverlap, Fs=fs)
-        if title is not None:
-            ax.set_title(title)
-        ax.set_ylim(0, 600)
-        ax.set_yticks(np.arange(0, 601, 50.0))
-        ax.set_xlabel("Time (sec)")
-        ax.set_ylabel("Frequencies (hz)")
-        plt.show()
-        plt.close()
-    data = decimate(data, 36)
-    spectrum, freqs, time, img = plt.specgram(data,  pad_to=nfft, NFFT=nfft, noverlap=noverlap, Fs=fs)
-    spectrum = abs(spectrum)
-    with open(name, 'wb') as outfile:
-        pickle.dump((spectrum, freqs, time), outfile)
+def show_spectrogram(data, date, file_time):
+    plt.specgram(data,  pad_to=nfft, NFFT=nfft, noverlap=noverlap, Fs=fs)
+    plt.title(date + "T" + file_time + "Z")
+    plt.ylim(0, 600)
+    plt.yticks(np.arange(0, 601, 50.0))
+    plt.xlabel("Time (sec)")
+    plt.ylabel("Frequencies (hz)")
+    plt.show()
     plt.close()
+
+
+def save_specgram_pkl(data, date, file_time, recording, output_dir, show=False):
+    data = resample(data, len(data) / 44100.0 * 2048)
+    if show:
+        show_spectrogram(data, date, file_time)
+    specgram, freqs, time, img = plt.specgram(data,  pad_to=nfft, NFFT=nfft, noverlap=noverlap, Fs=fs)
+
+    hex_num, hex_dir = Dates.to_hex(date, file_time)
+
+    specgram = abs(specgram)
+    print (date, file_time, hex_num)
+    print specgram.shape
+    for i in range(-1, specgram.shape[1] + 1):
+        temp_date, temp_time = Dates.add_seconds_to_date(date, file_time, i)
+        hex_num, hex_dir = Dates.to_hex(temp_date, temp_time)
+        if not os.path.isdir(output_dir + hex_dir):
+            os.makedirs(output_dir + hex_dir)
+        if "left" in recording:
+            output = output_dir + hex_dir + hex_num + "_" + temp_date + "T" + temp_time + "Z_left.spec.pkl"
+        else:
+            output = output_dir + hex_dir + hex_num + "_" + temp_date + "T" + temp_time + "Z_right.spec.pkl"
+        if i == -1:
+            spectrum = specgram[:, 0]
+        elif i == specgram.shape[1]:
+            spectrum = specgram[:, specgram.shape[1] - 1]
+        else:
+            spectrum = specgram[:, i]
+        if not os.path.isfile(output):
+            with open(output, 'wb') as outfile:
+                pickle.dump((spectrum, freqs, time), outfile)
+            plt.close()
 
 
 def main(input_dir, output_dir):
@@ -91,20 +113,10 @@ def main(input_dir, output_dir):
                     print (audio_dir + rec)
                     file_time = os.path.splitext(rec)[0][:os.path.splitext(rec)[0].index("_")]
                     date, file_time = Dates.convert_to_utc(date, file_time)
-                    hex_num, hex_dir = Dates.to_hex(date, file_time)
-                    if not os.path.isdir(output_dir + hex_dir):
-                        os.makedirs(output_dir + hex_dir)
-                    if "left" in rec:
-                        output = output_dir + hex_dir + hex_num + "_" + date + "T" + file_time + "Z_left.spec.pkl"
-                    else:
-                        output = output_dir + hex_dir + hex_num + "_" + date + "T" + file_time + "Z_right.spec.pkl"
-                    if not os.path.isfile(output):
-                        (bee_rate, bee_data) = get_data(audio_dir + rec)
-                        save_specgram_pkl(bee_data, os.path.splitext(rec)[0], output, show=False)
-                else:
-                    continue
+                    (bee_rate, bee_data) = get_data(audio_dir + rec)
+                    save_specgram_pkl(bee_data, date, file_time, rec, output_dir, show=False)
 
 
 if __name__ == "__main__":
     import sys
-    main(sys.argv[1], sys.argv[2])
+    main("/Users/lukestack/PycharmProjects/BeeVisualization", "/Users/lukestack/PycharmProjects/BeeVisualization/Pickles")
